Fix the three MSVC compile errors in the Windows build
- klsClipboard.h used std::string without including <string>
- A 64-bit wire ID narrowed into an unsigned long inside a braced list
- windows.h's min/max macros broke std::min/std::max in guiWire.cpp

+++ src/gui/GUICanvas.cpp
@@ -1362,7 +1362,7 @@ void GUICanvas::mouseRightDown(wxMouseEvent& event) {
 		// Part of a selection with other things in it: the menu acts on all of it.
 		const bool inSelection = std::find(selWires.begin(), selWires.end(), menuWire->getID()) != selWires.end()
 		                         && selWires.size() + selGates.size() > 1;
-		std::vector<unsigned long> targets = inSelection ? selWires : std::vector<unsigned long>{ menuWire->getID() };
+		std::vector<unsigned long> targets = inSelection ? selWires : std::vector<unsigned long>{ static_cast<unsigned long>(menuWire->getID()) };
 		if (inSelection) {
 			for (unsigned long id : selGates) if (guiGate* g = getGate(id)) g->select();
 			for (unsigned long id : selWires) if (guiWire* w = getWire(id)) w->select();
